+++ share/asciidoctor/convert_to_man.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+cd "$(dirname "$0")"
+
 version=$(go run ../../cmd/clx/main.go --version | awk '{print $NF}')
 
 touch clx.adoc